feat: Phase 4 - Integration and Import Updates (Issue #35)
Complete import updates and integration testing for the core restructure.

## Import Updates
- Updated all imports to use new module structure
- Fixed 20+ files using old core.framework imports
- Updated test files to use new paths
- Maintained all functionality

## Backward Compatibility
- Created comprehensive compatibility layers in core/framework/
- All old imports work with deprecation warnings
- Smooth migration path for existing code
- Zero breaking changes

## Import Mappings Applied
- core.framework.embedders → core.embedders
- core.framework.extractors → core.extractors
- core.framework.storage → core.workflows.storage
- core.processors.document_processor → core.workflows.workflow_pdf
- core.utils.state_manager → core.workflows.state

## Testing
- Created comprehensive integration test suite
- Verified new imports work correctly
- Confirmed backward compatibility
- All critical paths tested

This completes the integration phase of the core restructure,
ensuring all code works seamlessly with the new module structure.
