+++ tests/debug_segfault_isolated.py
@@ -19,7 +19,7 @@ import time
 project_root = Path(__file__).parent.parent
 sys.path.insert(0, str(project_root))
 
-from core.framework.extractors.docling_extractor import DoclingExtractor
+from core.extractors import DoclingExtractor
 
 # Setup detailed logging
 logging.basicConfig(
