+++ tests/test_config_files.py
@@ -10,7 +10,7 @@ import sys
 from pathlib import Path
 sys.path.insert(0, str(Path(__file__).parent.parent.parent))
 
-from core.framework.extractors.code_extractor import CodeExtractor
+from core.extractors import CodeExtractor
 import json
 import yaml
 import logging
